+++ app/helpers/projects_helper.rb
@@ -134,15 +134,34 @@ module ProjectsHelper
     options = project_feature_options
 
     if @project.private?
+      level = @project.project_feature.send(field)
       options.delete('Everyone with access')
-      highest_available_option = options.values.max if @project.project_feature.send(field) == ProjectFeature::ENABLED
+      highest_available_option = options.values.max if level == ProjectFeature::ENABLED
     end
 
     options = options_for_select(options, selected: highest_available_option || @project.project_feature.public_send(field))
-    content_tag(:select, options, name: "project[project_feature_attributes][#{field}]", id: "project_project_feature_attributes_#{field}", class: "pull-right form-control", data: { field: field }).html_safe
+
+    content_tag(
+      :select,
+      options,
+      name: "project[project_feature_attributes][#{field}]",
+      id: "project_project_feature_attributes_#{field}",
+      class: "pull-right form-control #{repo_children_classes(field)}",
+      data: { field: field }
+    ).html_safe
   end
 
   private
+
+  def repo_children_classes(field)
+    needs_repo_check = [:merge_requests_access_level, :builds_access_level]
+    return unless needs_repo_check.include?(field)
+
+    classes = "project-repo-select js-repo-select"
+    classes << " disabled" unless @project.feature_available?(:repository, current_user)
+
+    classes
+  end
 
   def get_project_nav_tabs(project, current_user)
     nav_tabs = [:home]
@@ -155,12 +174,8 @@ module ProjectsHelper
       nav_tabs << :merge_requests
     end
 
-    if can?(current_user, :read_pipeline, project)
-      nav_tabs << :pipelines
-    end
-
     if can?(current_user, :read_build, project)
-      nav_tabs << :builds
+      nav_tabs << :pipelines
     end
 
     if Gitlab.config.registry.enabled && can?(current_user, :read_container_image, project)
@@ -434,5 +449,9 @@ module ProjectsHelper
       'Only team members' => ProjectFeature::PRIVATE,
       'Everyone with access' => ProjectFeature::ENABLED
     }
+  end
+
+  def project_child_container_class(view_path)
+    view_path == "projects/issues/issues" ? "prepend-top-default" : "project-show-#{view_path}"
   end
 end